Extract CLI detection command execution
Make one runner own process startup, output capture, timeout cleanup, and exit
reporting while the detector keeps shell and version interpretation policy.
